feat: add diff filters and search for route tabs
