+++ frontend/src/pages/admin/routes/components/RouteGroupsTab.tsx
@@ -1,68 +1,207 @@
-import { DataTable } from '@/common/components';
+import { type ReactNode, useState } from 'react';
+
+import FilterLinesIcon from '@/assets/icons/filter-lines.svg?react';
+import {
+  Button,
+  DataTable,
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+  FilterChip,
+  FilterChipGroup,
+  SearchBar,
+} from '@/common/components';
 import type { Column } from '@/common/components';
 import type { RouteGroupRow } from '@/types/route-group';
 
 import { EmptyState } from './EmptyState';
 
+const WEEKDAYS = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri'];
+const DELIVERY_TYPES = ['School Year', 'Summer'];
+const ROUTE_STATUSES = ['Upcoming', 'Completed', 'Archived'];
+const DRIVER_STATUSES = ['Assigned', 'Unassigned'];
+
+interface FilterState {
+  weekdays: Set<string>;
+  deliveryTypes: Set<string>;
+  routeStatuses: Set<string>;
+  driverStatuses: Set<string>;
+}
+
+const emptyFilters = (): FilterState => ({
+  weekdays: new Set(),
+  deliveryTypes: new Set(),
+  routeStatuses: new Set(),
+  driverStatuses: new Set(),
+});
+
+const copyFilters = (f: FilterState): FilterState => ({
+  weekdays: new Set(f.weekdays),
+  deliveryTypes: new Set(f.deliveryTypes),
+  routeStatuses: new Set(f.routeStatuses),
+  driverStatuses: new Set(f.driverStatuses),
+});
+
 const COLUMNS: Column<RouteGroupRow>[] = [
-  {
-    key: 'name',
-    header: 'Name',
-    render: (row) => row.name,
-  },
-  {
-    key: 'date',
-    header: 'Date',
-    render: (row) => row.date,
-  },
+  { key: 'name', header: 'Name', render: (row) => row.name },
+  { key: 'date', header: 'Date', render: (row) => row.date },
   {
     key: 'delivery_type',
     header: 'Delivery Type',
     render: (row) => row.delivery_type,
   },
-  {
-    key: 'num_routes',
-    header: '# of Routes',
-    render: (row) => row.num_routes,
-  },
+  { key: 'num_routes', header: '# of Routes', render: (row) => row.num_routes },
   {
     key: 'num_locations',
     header: 'Locations',
     render: (row) => row.num_locations,
   },
-  {
-    key: 'num_boxes',
-    header: 'Boxes',
-    render: (row) => row.num_boxes,
-  },
+  { key: 'num_boxes', header: 'Boxes', render: (row) => row.num_boxes },
   {
     key: 'num_drivers_assigned',
     header: 'Drivers Assigned',
     render: (row) => row.num_drivers_assigned,
   },
-  {
-    key: 'status',
-    header: 'Status',
-    render: (row) => row.status,
-  },
+  { key: 'status', header: 'Status', render: (row) => row.status },
 ];
 
 interface RouteGroupsTabProps {
   rows?: RouteGroupRow[];
+  actions?: ReactNode;
 }
 
-export function RouteGroupsTab({ rows = [] }: RouteGroupsTabProps) {
+export function RouteGroupsTab({ rows = [], actions }: RouteGroupsTabProps) {
+  const [search, setSearch] = useState('');
+  const [filterOpen, setFilterOpen] = useState(false);
+  const [appliedFilters, setAppliedFilters] =
+    useState<FilterState>(emptyFilters());
+  const [draftFilters, setDraftFilters] = useState<FilterState>(emptyFilters());
+
+  const hasActiveFilters = Object.values(appliedFilters).some(
+    (s) => s.size > 0
+  );
+
+  const openFilters = () => {
+    setDraftFilters(copyFilters(appliedFilters));
+    setFilterOpen(true);
+  };
+
+  const toggleDraft = (key: keyof FilterState, value: string) => {
+    setDraftFilters((prev) => {
+      const next = new Set(prev[key]);
+      if (next.has(value)) next.delete(value);
+      else next.add(value);
+      return { ...prev, [key]: next };
+    });
+  };
+
+  const handleApply = () => {
+    setAppliedFilters(copyFilters(draftFilters));
+    setFilterOpen(false);
+    // TODO: pass appliedFilters to data fetching logic
+  };
+
   return (
-    <DataTable
-      columns={COLUMNS}
-      rows={rows}
-      getRowKey={(r) => r.id}
-      emptyState={
-        <EmptyState
-          title="No routes yet"
-          description="Try adjusting your filters or generating new routes"
-        />
-      }
-    />
+    <>
+      <div className="mb-8 flex items-center justify-between">
+        <div className="flex items-center gap-5">
+          <SearchBar
+            placeholder="Search anything"
+            value={search}
+            onChange={(e) => setSearch(e.target.value)}
+            wrapperClassName="w-64"
+          />
+          <Button
+            variant="tertiary"
+            shape="circular"
+            className={hasActiveFilters ? 'bg-blue-50' : 'bg-white'}
+            onClick={openFilters}
+          >
+            <FilterLinesIcon className="size-4" />
+          </Button>
+        </div>
+        {actions && <div className="flex items-center gap-4">{actions}</div>}
+      </div>
+
+      <DataTable
+        columns={COLUMNS}
+        rows={rows}
+        getRowKey={(r) => r.id}
+        emptyState={
+          <EmptyState
+            title="No routes yet"
+            description="Try adjusting your filters or generating new routes"
+          />
+        }
+      />
+
+      <Dialog open={filterOpen} onOpenChange={setFilterOpen}>
+        <DialogContent>
+          <DialogHeader>
+            <DialogTitle>Filters</DialogTitle>
+            <DialogDescription>Routes</DialogDescription>
+          </DialogHeader>
+
+          <div className="flex flex-col gap-4">
+            <FilterChipGroup label="Weekday">
+              {WEEKDAYS.map((day) => (
+                <FilterChip
+                  key={day}
+                  selected={draftFilters.weekdays.has(day)}
+                  onClick={() => toggleDraft('weekdays', day)}
+                >
+                  {day}
+                </FilterChip>
+              ))}
+            </FilterChipGroup>
+
+            <FilterChipGroup label="Delivery Type" showDelimiter>
+              {DELIVERY_TYPES.map((type) => (
+                <FilterChip
+                  key={type}
+                  selected={draftFilters.deliveryTypes.has(type)}
+                  onClick={() => toggleDraft('deliveryTypes', type)}
+                >
+                  {type}
+                </FilterChip>
+              ))}
+            </FilterChipGroup>
+
+            <FilterChipGroup label="Route Status" showDelimiter>
+              {ROUTE_STATUSES.map((status) => (
+                <FilterChip
+                  key={status}
+                  selected={draftFilters.routeStatuses.has(status)}
+                  onClick={() => toggleDraft('routeStatuses', status)}
+                >
+                  {status}
+                </FilterChip>
+              ))}
+            </FilterChipGroup>
+
+            <FilterChipGroup label="Driver Status" showDelimiter>
+              {DRIVER_STATUSES.map((status) => (
+                <FilterChip
+                  key={status}
+                  selected={draftFilters.driverStatuses.has(status)}
+                  onClick={() => toggleDraft('driverStatuses', status)}
+                >
+                  {status}
+                </FilterChip>
+              ))}
+            </FilterChipGroup>
+          </div>
+
+          <DialogFooter>
+            <Button variant="primary" onClick={handleApply}>
+              Apply
+            </Button>
+          </DialogFooter>
+        </DialogContent>
+      </Dialog>
+    </>
   );
 }
